+++ UnityProject/HoloCube/Assets/Scripts/ColorDetection/RgbEnums.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenCVForUnityExample.ColorDetection
+{
+    public enum RgbEnums
+    {
+        Red,
+        Green,
+        Blue
+    }
+}
