+++ UnityProject/HoloCube/Assets/Scripts/ColorDetection/RubiColors.cs
@@ -1,0 +1,13 @@
+﻿namespace OpenCVForUnityExample.ColorDetection
+{
+    public enum RubicColors
+    {
+        White,
+        Red,
+        Green,
+        Blue,
+        Yellow,
+        Orange,
+        Black
+    }
+}
